+++ src/Drastic.YouTube.Tests/TestData/VideoIds.cs
@@ -26,4 +26,5 @@ internal static class VideoIds
     public const string AgeRestrictedEmbedRestricted = "hySoCSoH-g8";
     public const string RatingDisabled = "5VGm0dczmHc";
     public const string RequiresPurchase = "p3dDcKOFXQg";
+    public const string HasChapters = "YQ_xWvX1n9g";
 }